+++ hw-05/src/main/java/ru/otus/booklibrary/dao/GenreDaoImpl.java
@@ -59,7 +59,7 @@ public class GenreDaoImpl implements GenreDao {
     public Genre insert(Genre genre) {
         BeanPropertySqlParameterSource parameterSource = new BeanPropertySqlParameterSource(genre);
         Number newKey = jdbcInsert.executeAndReturnKey(parameterSource);
-        genre.setId(newKey.intValue());
+        genre.setId(newKey.longValue());
         return genre;
     }
 }
